backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
gunicorn>=22.0
psycopg[binary]>=3.1
Pillow>=10.0
